Save progress

+++ app/controllers/api/lists_controller.rb
@@ -2,12 +2,8 @@ class Api::ListsController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    if current_user.id == params[:user_id].to_i
-      lists = User.find(params[:user_id]).lists
-      render json: lists, each_serializer: ListSerializer
-    else
-      render json: { message: "You're not authorized to see this list." }, status: :ok
-    end
+    lists = User.find(params[:user_id]).lists
+    render json: lists, each_serializer: ListSerializer
   end
 
   def create
@@ -22,22 +18,32 @@ class Api::ListsController < ApplicationController
   end
 
   def update
-    find_list
+    if user_authorized?
+      find_list
 
-    if @list.update(list_params)
-      render json: @list
+      if @list.update(list_params)
+        render json: @list
+      else
+        render json: { errors: list.errors.full_messages }, status: :unprocessable_entity
+      end
+
     else
-      render json: { errors: list.errors.full_messages }, status: :unprocessable_entity
+        not_authorized
     end
   end
 
   def destroy
-    find_list
+    if user_authorized?
+      find_list
 
-    if @list && @list.destroy
-      render json: { message: "List removed." }, status: :ok
+      if @list && @list.destroy
+        render json: { message: "List removed." }, status: :ok
+      else
+        record_not_found
+      end
+
     else
-      record_not_found
+        not_authorized
     end
   end
 
@@ -49,5 +55,9 @@ class Api::ListsController < ApplicationController
 
   def find_list
     @list = List.find_by_user_id_and_id(params[:user_id], params[:id])
+  end
+
+  def user_authorized?
+    current_user.id == params[:user_id].to_i
   end
 end
